+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.1.1 (2018-10-12)
+==================
+
+- use overlay\_config library
+- added --quiet argument, which suppress most output
+- fix generate\_json\_file
+- added --generate-report, which generates a report csv in output directory
+
 0.1.0 (2018-10-04)
 ==================